+++ src/gcovr/formats/html/__init__.py
@@ -20,7 +20,7 @@
 import logging
 from typing import List, Union
 
-from ...coverage import CovData
+from ...coverage import CoverageContainer
 from ...formats.base import BaseHandler
 from ...options import (
     GcovrConfigOption,
@@ -347,7 +347,7 @@ class HtmlHandler(BaseHandler):
         ):
             raise RuntimeError("only self contained reports can be printed to STDOUT")
 
-    def write_report(self, covdata: CovData, output_file: str) -> None:
+    def write_report(self, covdata: CoverageContainer, output_file: str) -> None:
         from .write import write_report  # pylint: disable=import-outside-toplevel # Lazy loading is intended here
 
         write_report(covdata, output_file, self.options)