+++ src/gcovr/formats/jacoco/__init__.py
@@ -19,7 +19,7 @@
 
 from typing import List, Union
 
-from ...coverage import CovData
+from ...coverage import CoverageContainer
 from ...formats.base import BaseHandler
 from ...options import GcovrConfigOption, OutputOrDefault
 
@@ -53,7 +53,7 @@ class JaCoCoHandler(BaseHandler):
             ),
         ]
 
-    def write_report(self, covdata: CovData, output_file: str) -> None:
+    def write_report(self, covdata: CoverageContainer, output_file: str) -> None:
         from .write import write_report  # pylint: disable=import-outside-toplevel # Lazy loading is intended here
 
         write_report(covdata, output_file, self.options)