+++ src/gcovr/formats/json/__init__.py
@@ -19,9 +19,9 @@
 
 import logging
 import os
-from typing import List, Optional, Union
+from typing import List, Union
 
-from ...coverage import CovData
+from ...coverage import CoverageContainer
 from ...formats.base import BaseHandler
 from ...options import GcovrConfigOption, OutputOrDefault
 from ...utils import force_unix_separator
@@ -111,17 +111,19 @@ class JsonHandler(BaseHandler):
             ),
         ]
 
-    def read_report(self) -> Optional[CovData]:
+    def read_report(self) -> CoverageContainer:
         from .read import read_report  # pylint: disable=import-outside-toplevel # Lazy loading is intended here
 
         return read_report(self.options)
 
-    def write_report(self, covdata: CovData, output_file: str) -> None:
+    def write_report(self, covdata: CoverageContainer, output_file: str) -> None:
         from .write import write_report  # pylint: disable=import-outside-toplevel # Lazy loading is intended here
 
         write_report(covdata, output_file, self.options)
 
-    def write_summary_report(self, covdata: CovData, output_file: str) -> None:
+    def write_summary_report(
+        self, covdata: CoverageContainer, output_file: str
+    ) -> None:
         from .write import write_summary_report  # pylint: disable=import-outside-toplevel # Lazy loading is intended here
 
         write_summary_report(covdata, output_file, self.options)