+++ src/gcovr/formats/txt/__init__.py
@@ -20,7 +20,7 @@
 import logging
 from typing import List, Union
 
-from ...coverage import CovData
+from ...coverage import CoverageContainer
 from ...formats.base import BaseHandler
 from ...options import (
     GcovrConfigOption,
@@ -102,12 +102,14 @@ class TxtHandler(BaseHandler):
             ),
         ]
 
-    def write_report(self, covdata: CovData, output_file: str) -> None:
+    def write_report(self, covdata: CoverageContainer, output_file: str) -> None:
         from .write import write_report  # pylint: disable=import-outside-toplevel # Lazy loading is intended here
 
         write_report(covdata, output_file, self.options)
 
-    def write_summary_report(self, covdata: CovData, output_file: str) -> None:
+    def write_summary_report(
+        self, covdata: CoverageContainer, output_file: str
+    ) -> None:
         from .write import write_summary_report  # pylint: disable=import-outside-toplevel # Lazy loading is intended here
 
         write_summary_report(covdata, output_file, self.options)