+++ CHANGELOG.rst
@@ -45,6 +45,7 @@ Internal changes:
 - Align variable names across the files. (:issue:`1015`)
 - Rework exclusion handling to keep information about excluded coverage. (:issue:`1016`)
 - Add ``mypy`` for testing code. (:issue:`1019`)
+- Add a container class for the coverage data. (:issue:`1023`)
 - Replace setup.py with hatchling. (:issue:`1026`)
 - Move gcovr to ``src`` directory. (:issue:`1027`)